+++ src/MineJason/TextChatComponent.cs
@@ -4,15 +4,20 @@
 
 namespace MineJason;
 
+using System.Text.Json.Serialization;
+using JetBrains.Annotations;
+
 /// <summary>
 /// Represents a chat component that displays a text.
 /// </summary>
 /// <param name="text">The text to display.</param>
+[PublicAPI]
 public sealed class TextChatComponent(string text) : ChatComponent("text"), IEquatable<TextChatComponent>
 {
     /// <summary>
     /// Gets the text to display.
     /// </summary>
+    [JsonPropertyName("text")]
     public string Text { get; } = text;
 
     /// <inheritdoc />